Added support for other System files
Because the coordinates and transparent color previously used were tuned for the default System file, the coordinates were changed to be more forgiving and the color is picked from the last pixel of the first row from the graphic.

Also, for some reason, every file in source control has been marked as modified, even though most haven't been. Don't know why this happened.