Engine: Imposed text color limit
Previously, if the client provided a message escape with an invalid color number (a.k.a. it wasn't between 0 and 19) FourPuzzle would crash with a confusing error message. This has been replaced with a more helpful message.

Also, previously, the colors between 10 and 19 could not be accessed, but this has been fixed.