Engine: Fixed walk-thru bug
Two specific events on the test map were always running over eachother. This problem started a few commits ago, but it was uncertain which commit this was. The key to solving the problem, as all other debugging didn't show anything, was that a certain event appeared to be blocked by something when it walked downward, which wasn't there. It was then discerned that that location was the original location of a moving event.

From there it was discerned that the parent map used by AbstractEvent was the original map that was copied from, not the map being used by MapViewGameState. A simple change to Map's copy() function fixed this.