Engine: Redesigned move frequency
Removed frequency from the Event API as LayerEvents are the only type of events that require a move frequency, as move frequency decides how many ticks to wait between each movement. HeroEvent does nothing with it, because the user input decides how often it moves.

Also changed frequency to be an integer rather than a bounded enumeration, which is clearer and lets an much larger (but probably unnecessary) range of wait values.