Engine: Fixed MessageWindow non-closure
Added code to MapViewGameState's deinitalize() that cancells the callbacks of all currently running events. Also added code to SpecialEvent's displayMessage() that allows an InterruptedException to kill the message.

With the problem with the Title Screen, it turned out that this bug had nothing to do with it, the Title Screen's interval was just too fast.

Fixes #18